Added binary search
iTunes.tracks() should return the list of tracks sorted alphabetically by artist, with some exceptions like 'The' or 'A', which are ignored by the sorting algorithm. With the newly added binary search it should be much quicker to find songs in such a list. Sorting exceptions are partly considered currently.